+++ app/tables/noteTable/elements/createAddNote.js
@@ -1,11 +1,11 @@
 import { createHTMLElem } from '../../../helpers/createHTMLElem.js';
-import { addNote } from '../addNote.js';
+import { addNoteDialog } from './dialog/addNoteDialog.js';
 
 export function createAddNote(parent$) {
 	const addNote$ = createHTMLElem(parent$, 'div', { class: 'add-note' });
 
 	const addNoteButton$ = createHTMLElem(addNote$, 'button', { class: 'add-note__button' }, '➕ Add note');
-	addNoteButton$.addEventListener('click', addNote);
+	addNoteButton$.addEventListener('click', addNoteDialog);
 
 	return parent$;
 }
