+++ app/tables/noteTable/elements/dialog/addNoteDialog.js
@@ -1,7 +1,7 @@
-import { store } from '../../store/store.js';
-import { createDialog } from './elements/dialog/createDialog.js';
+import { store } from '../../../../store/store.js';
+import { createDialog } from './createDialog.js';
 
-export function addNote() {
+export function addNoteDialog() {
 	createDialog('ADD NOTE', '', '', '', (data) => {
 		const { name, category, content } = data;
 		store.data.addNote(name, category, content);
